pyproject.toml
[tool.poetry]
name = "NanoAODProdTool"
version = "0.1.0"
description = ""
authors = ["Dietrich Liko <Dietrich.Liko@oeaw.ac.at>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
Jinja2 = "^3.1.2"
finite-state-machine = "^0.5.0"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
mypy = "^0.960"
flake8 = "^4.0.1"
flake8-bugbear = "^22.4.25"
reorder-python-imports = "^3.1.0"

[tool.poetry.scripts]
nanojobs = "nanoaodprodtool.cli:cli"
nanojobs_meta =  "nanoaodprodtool.meta:main"
nanojobs_test = "nanoaodprodtool.test:main"
list_datasets = "nanoaodprodtool.list_dsets:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
